chore: move get-events to actions
