+++ src/app/api/events/repository.ts
@@ -1,4 +1,4 @@
-import { Event } from '@/app/types/event';
+import { Event } from '@/types/event';
 import { PrismaClient } from '@/generated/client';
 import { EventBody } from './route';
