--- src/app/service.ts
@@ -1,10 +1,0 @@
-'use server';
-
-import { Event } from './types/event';
-import { getAllEvents as getAllEventsFromService } from '@/app/api/events/service';
-
-export async function getAllEvents(): Promise<Event[]> {
-  const events = await getAllEventsFromService();
-
-  return events;
-}
