+++ src/components/calendar/calendar.tsx
@@ -1,4 +1,4 @@
-import { Event } from '@/app/types/event';
+import { Event } from '@/types/event';
 import { EventComponent } from '../event/event-component';
 
 type CalendarProps = {
